+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/downloader/ManagedDownloader.java
@@ -542,7 +542,7 @@ public class ManagedDownloader extends AbstractDownloader
 		if (propertiesMap.get(DEFAULT_FILENAME) == null)
 			propertiesMap.put(DEFAULT_FILENAME,rfd.getFileName());
 		if (propertiesMap.get(FILE_SIZE) == null)
-			propertiesMap.put(FILE_SIZE,new Integer(rfd.getSize()));
+			propertiesMap.put(FILE_SIZE,new Long(rfd.getSize()));
     }
     
     /** 
@@ -1326,7 +1326,7 @@ public class ManagedDownloader extends AbstractDownloader
 	 * @param fileSize, can be 0
 	 * @return
 	 */
-	public boolean conflicts(URN urn, int fileSize, File... fileName) {
+	public boolean conflicts(URN urn, long fileSize, File... fileName) {
 		if (urn != null && downloadSHA1 != null) {
 			return urn.equals(downloadSHA1);
 		}
